Updated MeasurementsDTO, converted to BigDecimal

+++ src/main/java/pl/maciejak/my_portfolio_rest/dto/MeasurementsDTO.java
@@ -1,9 +1,10 @@
 package pl.maciejak.my_portfolio_rest.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record MeasurementsDTO(Double productLength,
-                              Double negTolerance,
-                              Double posTolerance,
-                              List<Double> measurements) {
+public record MeasurementsDTO(BigDecimal productLength,
+                              BigDecimal negTolerance,
+                              BigDecimal posTolerance,
+                              List<BigDecimal> measurements) {
 }
